Fix string initializer for ShellNavigationState (#13478) fixes #13422
