+++ Xamarin.Forms.Core.UnitTests/ShellNavigationStateTests.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+using NUnit.Framework;
+using Xamarin.Forms.Internals;
+
+namespace Xamarin.Forms.Core.UnitTests
+{
+	[TestFixture]
+	public class ShellNavigationStateTests : ShellTestBase
+	{
+		[Test]
+		public void LocationInitializedWithUri()
+		{
+			var uri = new Uri($"//test/IMPL_TEST/D_FAULT_TEST", UriKind.Relative);
+			var uriState = new ShellNavigationState(uri);
+
+			Assert.AreEqual("//test", uriState.Location.ToString());
+			Assert.AreEqual("//test/IMPL_TEST/D_FAULT_TEST", uriState.FullLocation.ToString());
+		}
+
+		[Test]
+		public void LocationInitializedWithString()
+		{
+			var uri = new Uri("//test/IMPL_TEST/D_FAULT_TEST", UriKind.Relative);
+			var strState = new ShellNavigationState(uri.ToString());
+
+			Assert.AreEqual("//test", strState.Location.ToString());
+			Assert.AreEqual("//test/IMPL_TEST/D_FAULT_TEST", strState.FullLocation.ToString());
+		}
+	}
+}
